added stuff

+++ transformer/myvit.py
@@ -62,9 +62,9 @@ data_resize = tf.keras.Sequential(
         )
 data_resize.layers[0].adapt(x_test)
 
-train_dataset = tf.cast(tf.data.Dataset.from_tensor_slices((x_train, y_train)),tf.float32)
+train_dataset = tf.data.Dataset.from_tensor_slices((x_train, y_train))
 train_dataset = train_dataset.batch(batch_size).map(lambda x, y: (data_resize_aug(x), y))
-test_dataset = tf.cast(tf.data.Dataset.from_tensor_slices((x_train, y_train)),tf.float32)
+test_dataset = tf.data.Dataset.from_tensor_slices((x_train, y_train))
 test_dataset = test_dataset.batch(batch_size).map(lambda x, y: (data_resize(x), y))
 def run_experiment(model):
